% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_seirdv.R
\name{build_seirdv}
\alias{build_seirdv}
\title{Function to create a SEIRDV model}
\usage{
build_seirdv()
}
\description{
This function creates an odin instance of the stochastic, discrete SEIRDV
model. Since parameter inputs can be specified at runtime at no cost, this
function is used for internal purposes only. See \code{run_seirdv()} for running
simulations using this model.
}
\author{
Thibaut Jombart
}
