% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_inputs.R
\name{process_inputs}
\alias{process_inputs}
\title{Internal function to check inputs}
\usage{
process_inputs(x)
}
\arguments{
\item{x}{A named list of inputs for run_seirdv()}
}
\description{
The functions \code{check_...} return nothing but issue errors as needed.
The functions \code{process_...} return their input, modified if needed, and
issue errors as needed.
}
\author{
Thibaut Jombart
}
